// src/user/interfaces/profile.interface.ts
export interface iProfile {
    firstName: string
    lastName: string
    age: number
    phone: string
    city: string
    country: string
    postalCode: string
    address: string
    gender: string
    dateOfBirth: Date
    avatarUrl: string
    bio: string
    socialMediaLinks: string
    interests: string
    education: string
    employment: string
}